Split oversized chat settings components
